+++ lib/order_query/sql/where.rb
@@ -8,12 +8,12 @@ module OrderQuery
       # @param [OrderQuery::Point] point
       def initialize(point)
         @point      = point
-        @conditions = point.space.conditions
+        @columns = point.space.columns
       end
 
-      # Join condition pairs with OR, and nest within each other with AND
+      # Join column pairs with OR, and nest within each other with AND
       # @param [:before or :after] side
-      # @return [query, parameters] WHERE conditions matching records strictly before / after this one
+      # @return [query, parameters] WHERE columns matching records strictly before / after this one
       #   sales < 5 OR
       #   sales = 5 AND (
       #     invoice < 3 OR
@@ -21,14 +21,14 @@ module OrderQuery
       #       ... ))
       def build(side)
         # generate pairs of terms such as sales < 5, sales = 5
-        parts = @conditions.map { |cond|
-          [where_side(cond, side, true), where_tie(cond)].reject { |x| x == WHERE_IDENTITY }
+        terms = @columns.map { |col|
+          [where_side(col, side, true), where_tie(col)].reject { |x| x == WHERE_IDENTITY }
         }
         # group pairwise with OR, and nest with AND
-        query = foldr_terms parts.map { |terms| join_terms 'OR'.freeze, *terms }, 'AND'.freeze
+        query = foldr_terms terms.map { |pair| join_terms 'OR'.freeze, *pair }, 'AND'.freeze
         if ::OrderQuery.wrap_top_level_or
           # wrap in a redundant AND clause for performance
-          query = wrap_top_level_or query, parts, side
+          query = wrap_top_level_or query, terms, side
         end
         query
       end
@@ -44,7 +44,7 @@ module OrderQuery
         end
       end
 
-      # joins terms with an operator
+      # joins terms with an operator, empty terms are skipped
       # @return [query, parameters]
       def join_terms(op, *terms)
         [terms.map(&:first).reject(&:empty?).join(" #{op} "), terms.map(&:second).reduce([], :+)]
@@ -63,61 +63,55 @@ module OrderQuery
       #    (sales < 5 OR
       #       (sales = 5 AND ...)))
       # Read more at https://github.com/glebm/order_query/issues/3
-      def wrap_top_level_or(query, pairs, side)
-        top_pair_idx = pairs.index(&:present?)
-        if top_pair_idx &&
-            (top_pair = pairs[top_pair_idx]).length == 2 &&
-            top_pair.first != (redundant_cond = where_side(@conditions[top_pair_idx], side, false))
-          join_terms 'AND'.freeze, redundant_cond, wrap_term_with_parens(query)
+      def wrap_top_level_or(query, terms, side)
+        top_term_i = terms.index(&:present?)
+        if top_term_i && terms[top_term_i].length == 2 &&
+            (cond = where_side(@columns[top_term_i], side, false)) != WHERE_IDENTITY
+          join_terms 'AND'.freeze, cond, wrap_term_with_parens(query)
         else
           query
         end
       end
 
-      # @return [query, params] tie-breaker unless condition is unique
-      def where_tie(cond)
-        if cond.unique?
+      # @return [query, params] tie-breaker unless column is unique
+      def where_tie(col)
+        if col.unique?
           WHERE_IDENTITY
         else
-          where_eq(cond)
+          where_eq(col)
         end
       end
 
       # @param [:before or :after] side
-      # @return [query, params] return query conditions for attribute values before / after the current one
-      def where_side(cond, side, strict = true, value = point.value(cond))
-        if cond.order_enum
-          values = cond.enum_side(value, side, strict)
-          if cond.complete? && values.length == cond.order_enum.length
-            WHERE_IDENTITY
-          else
-            where_in cond, values
-          end
+      # @return [query, params] return query fragment for column values before / after the current one
+      def where_side(col, side, strict = true, value = point.value(col))
+        if col.order_enum
+          where_in col, col.enum_side(value, side, strict)
         else
-          where_ray cond, value, side, strict
+          where_ray col, value, side, strict
         end
       end
 
-      def where_in(cond, values)
+      def where_in(col, values)
         case values.length
           when 0
             WHERE_IDENTITY
           when 1
-            where_eq cond, values[0]
+            where_eq col, values[0]
           else
-            ["#{cond.column_name} IN (?)".freeze, [values]]
+            ["#{col.column_name} IN (?)".freeze, [values]]
         end
       end
 
-      def where_eq(cond, value = point.value(cond))
-        [%Q(#{cond.column_name} = ?).freeze, [value]]
+      def where_eq(col, value = point.value(col))
+        [%Q(#{col.column_name} = ?).freeze, [value]]
       end
 
-      def where_ray(cond, from, mode, strict = true)
+      def where_ray(col, from, mode, strict = true)
         ops = %w(< >)
         ops = ops.reverse if mode == :after
-        op  = {asc: ops[0], desc: ops[1]}[cond.order || :asc]
-        ["#{cond.column_name} #{op}#{'=' unless strict} ?".freeze, [from]]
+        op  = {asc: ops[0], desc: ops[1]}[col.order || :asc]
+        ["#{col.column_name} #{op}#{'=' unless strict} ?".freeze, [from]]
       end
 
       WHERE_IDENTITY = [''.freeze, [].freeze].freeze